backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
openai
psycopg2-binary
